+++ src/lib.rs
@@ -1,13 +1,21 @@
 #![doc = include_str!("../README.md")]
 
+/// Internal default panic message used by Empty variant types.
+pub(crate) const DEFAULT_DROP_PANIC_MSG: &str = "Value was dropped without being unwrapped";
+
 #[warn(clippy::pedantic)]
 #[warn(clippy::nursery)]
 #[warn(clippy::cargo)]
 #[warn(missing_docs)]
+#[warn(missing_debug_implementations)]
 #[allow(clippy::match_bool)]
-mod guards;
-mod into;
-mod markers;
+#[allow(clippy::single_match_else)]
+#[doc(hidden)]
+pub mod guards;
+#[doc(hidden)]
+pub mod into;
+#[doc(hidden)]
+pub mod markers;
 
 #[doc(hidden)]
 pub mod wrap;
@@ -16,13 +24,12 @@ pub mod wrap;
 /// with debug-only panic behavior.
 pub mod dbg {
     pub use crate::guards::GuardNotArmed;
-    pub use crate::wrap::DEFAULT_DROP_PANIC_MSG;
 
     #[cfg(debug_assertions)]
     pub use crate::wrap::NoDrop;
 
     #[cfg(not(debug_assertions))]
-    pub use crate::wrap::NoDropPassEmpty as NoDropEmpty;
+    pub use crate::wrap::NoDropNoOpEmpty as NoDropEmpty;
 
     #[cfg(debug_assertions)]
     pub use crate::into::IntoNoDropRls as IntoNoDrop;
@@ -34,25 +41,24 @@ pub mod dbg {
     pub use crate::wrap::NoDropMsg;
 
     #[cfg(not(debug_assertions))]
-    pub use crate::wrap::NoDropPassMsg as NoDropMsg;
+    pub use crate::wrap::NoDropNoOpMsg as NoDropMsg;
 
     #[cfg(debug_assertions)]
     pub use crate::guards::DropGuardEmpty;
 
     #[cfg(not(debug_assertions))]
-    pub use crate::guards::DropGuardPassthroughEmpty as DropGuardEmpty;
+    pub use crate::guards::DropGuardNoOpEmpty as DropGuardEmpty;
 
     #[cfg(debug_assertions)]
     pub use crate::guards::DropGuard;
 
     #[cfg(not(debug_assertions))]
-    pub use crate::guards::DropGuardPassthroughMsg as DropGuard;
+    pub use crate::guards::DropGuardNoOpMsg as DropGuard;
 }
 
-/// Module containing [`NoDrop`](no_drop::NoDropEmpty) and [`NoDropMsg`](no_drop::NoDropMsg) with always-[`panic!`]ing behavior.
+/// Module containing [`NoDrop`](wrap::NoDropEmpty) and [`NoDropMsg`](wrap::NoDropMsg) with always-[`panic!`]ing behavior.
 pub mod rls {
     pub use crate::guards::GuardNotArmed;
-    pub use crate::wrap::DEFAULT_DROP_PANIC_MSG;
 
     pub use crate::wrap::NoDrop;