+++ src/wrap/mod.rs
@@ -1,15 +1,16 @@
-#[cfg(test)]
-mod test_macros;
-
-mod no_drop_empty;
-mod no_drop_msg;
-mod no_drop_pass;
+#[doc(hidden)]
+pub mod no_drop_empty;
+#[doc(hidden)]
+pub mod no_drop_msg;
+#[doc(hidden)]
+pub mod no_drop_no_op_empty;
+#[doc(hidden)]
+pub mod no_drop_no_op_msg;
 
 pub use no_drop_empty::NoDropEmpty as NoDrop;
-pub(crate) use no_drop_empty::NoDropEmpty;
-pub use no_drop_empty::DEFAULT_DROP_PANIC_MSG;
+pub use no_drop_empty::NoDropEmpty;
 pub use no_drop_msg::NoDropMsg;
 
-// Type aliases for passthrough variants
-pub type NoDropPassEmpty<T = ()> = no_drop_pass::NoDropPass<'static, crate::markers::Empty, T>;
-pub type NoDropPassMsg<'msg, T = ()> = no_drop_pass::NoDropPass<'msg, crate::markers::Msg, T>;
+// Type aliases for no_op variants
+pub type NoDropNoOpEmpty<T = ()> = no_drop_no_op_empty::NoDropNoOp<T>;
+pub type NoDropNoOpMsg<'msg, T = ()> = no_drop_no_op_msg::NoDropNoOp<'msg, T>;